refactor(components): abstract InferDialog
