+++ src/app/components/SubmitButton.tsx
@@ -1,12 +1,18 @@
 import { Button } from "@/components/ui/button";
 import { useFormStatus } from "react-dom";
 
-export const SubmitButton = () => {
+export const SubmitButton = ({
+  pendingLabel,
+  submitLabel,
+}: {
+  pendingLabel: string;
+  submitLabel: string;
+}) => {
   const { pending } = useFormStatus();
 
   return (
     <Button type="submit" disabled={pending}>
-      {pending ? "Infering issues..." : "Infer"}
+      {pending ? pendingLabel : submitLabel}
     </Button>
   );
 };
